test: remove global pollution
